feat(trie): stored nibbles
