+++ crates/primitives/src/trie/mod.rs
@@ -1,7 +1,7 @@
 //! Collection of trie related types.
 
 mod nibbles;
-pub use nibbles::Nibbles;
+pub use nibbles::{StoredNibbles, StoredNibblesSubKey};
 
 mod branch_node;
 pub use branch_node::BranchNodeCompact;
